pyproject.toml
[project]
name = "firefly-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.64.0",
    "pyinstaller>=6.12.0",
    "python-dotenv>=1.0.1",
    "pywebview>=5.4",
]
